feat: change edit profile logic

+++ src/infrastructure/storage/gcsStorage.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { Storage } from '@google-cloud/storage';
 import { IGcsStorage } from 'src/domains/repositories/storage/IGcsStorage';
 
@@ -34,9 +34,24 @@ export class GcsStorageService implements IGcsStorage {
     return `https://storage.googleapis.com/${this.bucketName}/${destination}`;
   }
   async deleteFile(filePath: string): Promise<void> {
+    // Decode URL encoded characters
+    const decodedPath = decodeURIComponent(filePath);
+    
     const bucket = this.storage.bucket(this.bucketName);
-    const file = bucket.file(filePath);
-
-    await file.delete();
+    const file = bucket.file(decodedPath);
+  
+    try {
+      // Check if file exists before attempting to delete
+      const [exists] = await file.exists();
+      
+      if (exists) {
+        await file.delete();
+      } else {
+        throw new NotFoundException(`File ${decodedPath} not found`);
+      }
+    } catch (error) {
+      console.log(error);
+      throw new BadRequestException(`Failed to delete file: ${decodedPath}`);
+    }
   }
 }
